backend/requirements.txt
fastapi
uvicorn[standard]
numpy
nibabel
nilearn
mne
matplotlib
pillow
scikit-learn
pandas
python-multipart
